feat(inventory groups): fixed pr after review
