+++ src/components/InventoryGroups/Modals/ModalSchemas/schemes.js
@@ -1,6 +1,8 @@
+import React from 'react';
 import validatorTypes from '@data-driven-forms/react-form-renderer/validator-types';
 import componentTypes from '@data-driven-forms/react-form-renderer/component-types';
 import { nameValidator } from '../../helpers/validate';
+import { Text } from '@patternfly/react-core';
 
 export const createGroupSchema = (namePresenceValidator) => ({
     fields: [
@@ -20,5 +22,34 @@ export const createGroupSchema = (namePresenceValidator) => ({
                 nameValidator
             ]
         }
+    ]
+});
+
+const createDescription = (systemName) => {
+    return (
+        <Text>
+        Select a group to add <strong>{systemName}</strong> to, or create a new one.
+        </Text>
+    );
+};
+
+//this is a custom schema that is passed via additional mappers to the Modal component
+//it allows to create custom item types in the modal
+
+export const addHostSchema = (systemName) => ({
+    fields: [
+        {
+            component: componentTypes.PLAIN_TEXT,
+            name: 'description',
+            label: createDescription(systemName)
+        },
+        {
+            component: 'search-input',
+            name: 'group',
+            label: 'Select a group',
+            isRequired: true,
+            validate: [{ type: validatorTypes.REQUIRED }]
+        },
+        { component: 'create-group-btn', name: 'create-group-btn' }
     ]
 });
